requirements.txt
aiosqlite==0.16.1
astroid==2.4.2
asyncpg==0.21.0
click==7.1.2
dataclasses==0.6
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
iso8601==0.1.13
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pydantic==1.7.3
pylint==2.6.0
PyPika==0.44.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.4.1
six==1.15.0
SQLAlchemy==1.3.22
starlette==0.13.6
toml==0.10.2
tortoise-orm==0.16.20
typed-ast==1.4.2
typing-extensions==3.7.4.3
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
wrapt==1.12.1
